+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.44.1] - 2026-04-16
+
 ### Fixed
 
 - Fixed event migration failure introduced in 0.44.0 where
@@ -1390,7 +1392,8 @@ AsRef<[u8]>`). This change accommodates scenarios where the information stored
 - Modified `FtpBruteForce` by adding an `is_internal` field which is a boolean
   indicating whether it is internal or not.
 
-[Unreleased]: https://github.com/petabi/review-database/compare/0.44.0...main
+[Unreleased]: https://github.com/petabi/review-database/compare/0.44.1...main
+[0.44.1]: https://github.com/petabi/review-database/compare/0.44.0...0.44.1
 [0.44.0]: https://github.com/petabi/review-database/compare/0.43.0...0.44.0
 [0.43.0]: https://github.com/petabi/review-database/compare/0.42.0...0.43.0
 [0.42.0]: https://github.com/petabi/review-database/compare/0.41.0...0.42.0